Removed crop to ecoregion process used for the HHM simple engine
